requirements.txt
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
fpdf2==2.5.6
greenlet==1.1.2
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.0.1
Jinja2==3.1.2
MarkupSafe==2.1.1
Pillow==9.2.0
pycparser==2.21
SQLAlchemy==1.4.39
svg.path==6.2
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.1
